fix: :bug: Solved "fetchTranslations executed at last"
I have changed the strategy to fetchTranslations. It used to be called from ResponsePrepared event but it appears to be that event is called after everything is done. And this causes library to be unable to find translations even if they were found. I moved fetchTranslations right into translate function and checked if thats the first attempt to retreive a cached translation. If its first, it calls the fetchTranslations.